requirements.txt
certifi==2019.3.9
Django==2.1.7
image==1.5.27
MIDIUtil==1.2.1
Pillow==5.4.1
pytz==2018.9
